feat: Actions for api, leaving the workplace funcionality

+++ app/routes/api+/workplace+/index.tsx
@@ -1,33 +1,43 @@
-import { type ActionFunctionArgs } from '@remix-run/node';
+import { redirect, type ActionFunctionArgs } from '@remix-run/node';
 import { eq } from 'drizzle-orm';
+import { $path } from 'remix-routes';
 import { z } from 'zod';
 import { zx } from 'zodix';
 
-import { db, deleteDatabase } from '~/db';
+import { db, deleteDatabase, workplaceDb } from '~/db';
 import { workplaceMemberTable, workplaceTable } from '~/db/schema';
-import { projectTable } from '~/db/schema-workplace';
+import { workplaceUser } from '~/db/schema-workplace';
 import { requireUser } from '~/services/auth.server';
 
 export async function action({ request, params }: ActionFunctionArgs) {
   const user = await requireUser({ request, params });
 
-  const { workplaceId } = await zx.parseForm(request, {
-    workplaceId: z.string().min(1)
+  const { workplaceId, _action } = await zx.parseForm(request, {
+    workplaceId: z.string().min(1),
+    _action: z.enum(['delete', 'leave'])
   });
 
-  const ownerWorkplaces = user.ownerOfWorkplace.map((i) => i.id);
-  if (!ownerWorkplaces.includes(workplaceId)) {
-    throw Error('You are not the owner of this workplace');
+  if (_action === 'delete') {
+    const ownerWorkplaces = user.ownerOfWorkplace.map((i) => i.id);
+    if (!ownerWorkplaces.includes(workplaceId)) {
+      throw Error('You are not the owner of this workplace');
+    }
+    // //Remove Workplace members
+    await db.delete(workplaceMemberTable).where(eq(workplaceMemberTable.workplaceId, workplaceId));
+
+    // // Remove Workplace
+    await db.delete(workplaceTable).where(eq(workplaceTable.id, workplaceId));
+
+    //Remove Db
+    return await deleteDatabase(workplaceId);
   }
-  //Remove Workplace members
-  await db.delete(workplaceMemberTable).where(eq(workplaceMemberTable.workplaceId, workplaceId));
+  if (_action === 'leave') {
+    //Remove Workplace member
+    await db.delete(workplaceMemberTable).where(eq(workplaceMemberTable.userId, user.id));
 
-  // remove Project
-  await db.delete(projectTable);
+    //Remove myself from the workplace
+    await workplaceDb(workplaceId).delete(workplaceUser).where(eq(workplaceUser.id, user.id));
 
-  // Remove Workplace
-  await db.delete(workplaceTable).where(eq(workplaceTable.id, workplaceId));
-
-  //Remove Db
-  return await deleteDatabase(workplaceId);
+    return redirect($path('/app'));
+  }
 }
